Add a mixture component
- separate mixture component from IFU
- allowed for deletion of lots of duplication between the two
- adjust tests and docstrings accordingly
