+++ popkinmocks/components/__init__.py
@@ -1,3 +1,4 @@
 from .growing_disk import GrowingDisk
 from .stream import Stream
 from .base import Component
+from .mixture import Mixture
